+++ mapchete/__init__.py
@@ -585,14 +585,14 @@ class MapcheteProcess(object):
             else:
                 return existing_tile.read(**kwargs)
 
-    def open(self, input_file, **kwargs):
+    def open(self, input_id, **kwargs):
         """
         Open input data.
 
         Parameters
         ----------
-        input_file : string
-            file identifier from configuration file or file path
+        input_id : string
+            input identifier from configuration file or file path
         kwargs : driver specific parameters (e.g. resampling)
 
         Returns
@@ -600,12 +600,12 @@ class MapcheteProcess(object):
         tiled input data : InputTile
             reprojected input data within tile
         """
-        if not isinstance(input_file, str):
-            return input_file.open(self.tile, **kwargs)
-        if input_file not in self.params["input_files"]:
+        if not isinstance(input_id, str):
+            return input_id.open(self.tile, **kwargs)
+        if input_id not in self.params["input"]:
             raise ValueError(
-                "%s not found in config as input file" % input_file)
-        return self.params["input_files"][input_file].open(self.tile, **kwargs)
+                "%s not found in config as input file" % input_id)
+        return self.params["input"][input_id].open(self.tile, **kwargs)
 
     def hillshade(
         self, elevation, azimuth=315.0, altitude=45.0, z=1.0, scale=1.0