+++ mapchete/config/__init__.py
@@ -13,13 +13,16 @@ when initializing the configuration.
 import os
 import yaml
 import logging
+import warnings
 from cached_property import cached_property
 from shapely.geometry import box, MultiPolygon
 
-from mapchete.formats import load_output_writer, available_output_formats
+from mapchete.formats import (
+    load_output_writer, available_output_formats, available_input_formats
+)
 from mapchete.tile import BufferedTilePyramid
 from mapchete.errors import MapcheteConfigError
-from mapchete.config._parse_input_files import input_files_at_zoom
+from mapchete.config._parse_input import input_at_zoom
 
 
 LOGGER = logging.getLogger(__name__)
@@ -29,20 +32,20 @@ TILING_TYPES = ["geodetic", "mercator"]
 
 # parameters to be provided in the process configuration
 _MANDATORY_PARAMETERS = [
-    "process_file",  # the Python file the process is defined in
-    "input_files",  # input files for process; can also be "from_command_line"
-    "output"  # dictionary configuring the output format
+    "process_file",     # the Python file the process is defined in
+    "input",            # files & other types; can also be "from_command_line"
+    "output"            # process output format parameters
 ]
 
 # parameters with special functions which cannot be used for user parameters
 _RESERVED_PARAMETERS = [
     "process_minzoom",  # minimum zoom where process is valid
     "process_maxzoom",  # maximum zoom where process is valid
-    "process_zoom",  # single zoom where process is valid
-    "process_bounds",  # process boundaries
-    "metatiling",  # metatile size (for both process and output)
-    "pixelbuffer",  # buffer around each tile in pixels
-    "baselevels"  # enable interpolation from other zoom levels
+    "process_zoom",     # single zoom where process is valid
+    "process_bounds",   # process boundaries
+    "metatiling",       # process metatile size
+    "pixelbuffer",      # buffer around each tile in pixels
+    "baselevels"        # enable interpolation from other zoom levels
 ]
 
 
@@ -112,19 +115,18 @@ class MapcheteConfig(object):
         LOGGER.info("preparing configuration ...")
         if debug:
             LOGGER.setLevel(logging.DEBUG)
-        try:
-            assert mode in ["memory", "readonly", "continue", "overwrite"]
-        except Exception:
+        if not mode in ["memory", "readonly", "continue", "overwrite"]:
             raise MapcheteConfigError("invalid process mode")
         LOGGER.debug("zooms provided to config: %s" % zoom)
         self.mode = mode
         # parse configuration
         LOGGER.debug("parse configuration ...")
-        self._input_files_cache = {}
+        self._input_cache = {}
         self._process_area_cache = {}
         self.raw, self.mapchete_file, self.config_dir = self._parse_config(
             input_config, single_input_file=single_input_file)
-        assert self.process_file
+        if not self.process_file:
+            raise MapcheteConfigError("no process_file given")
         # set process delimiters
         self._delimiters = dict(zoom=zoom, bounds=bounds)
         # helper caches
@@ -135,9 +137,7 @@ class MapcheteConfig(object):
             self.output_type = self.raw["output"]["type"]
         except KeyError:
             raise MapcheteConfigError("no output type given")
-        try:
-            assert self.raw["output"]["type"] in TILING_TYPES
-        except AssertionError:
+        if self.raw["output"]["type"] not in TILING_TYPES:
             raise MapcheteConfigError(
                 "output type (geodetic or mercator) is missing")
         self.process_pyramid = BufferedTilePyramid(
@@ -162,23 +162,21 @@ class MapcheteConfig(object):
                     output_params["format"], str(available_output_formats())
                 ))
         writer = load_output_writer(output_params)
-        try:
-            assert writer.is_valid_with_config(output_params)
-        except AssertionError, e:
-            raise ValueError(
-                "driver %s not compatible with configuration: %s" %
-                (writer.METADATA["driver_name"], e))
+        if not writer.is_valid_with_config(output_params):
+            raise MapcheteConfigError(
+                "driver %s not compatible with configuration: %s" % (
+                    writer.METADATA["driver_name"], output_params)
+                )
         return writer
 
     @cached_property
     def process_file(self):
         """Absolute path of process file."""
         abs_path = os.path.join(self.config_dir, self.raw["process_file"])
-        try:
-            assert os.path.isfile(abs_path)
-        except Exception:
+        if os.path.isfile(abs_path):
+            return abs_path
+        else:
             raise MapcheteConfigError("%s is not available" % abs_path)
-        return abs_path
 
     @cached_property
     def zoom_levels(self):
@@ -201,16 +199,12 @@ class MapcheteConfig(object):
         if isinstance(zoom, int):
             zoom = [zoom]
         if len(zoom) == 1:
-            try:
-                assert zoom[0] >= 0
-            except Exception:
+            if zoom[0] < 0:
                 raise MapcheteConfigError("Zoom level must be greater 0.")
             return zoom
         elif len(zoom) == 2:
             for i in zoom:
-                try:
-                    assert i >= 0
-                except Exception:
+                if i < 0:
                     raise MapcheteConfigError("Zoom levels must be greater 0.")
             if zoom[0] < zoom[1]:
                 return range(zoom[0], zoom[1]+1)
@@ -235,34 +229,28 @@ class MapcheteConfig(object):
             baselevels = self.raw["baselevels"]
         except KeyError:
             return {}
-        try:
-            minmax = {
-                k: v for k, v in baselevels.iteritems() if k in ["min", "max"]
-            }
-            assert minmax
-            for v in minmax.values():
-                assert v >= 0
-                assert isinstance(v, int)
-        except Exception as e:
+        minmax = {
+            k: v for k, v in baselevels.iteritems() if k in ["min", "max"]
+        }
+        if not minmax:
             raise MapcheteConfigError(
-                "invalid baselevel zoom parameter given: %s" % e
+                "no min and max values given for baselevels"
             )
-        try:
-            base_min = minmax["min"]
-        except KeyError:
-            base_min = min(self.zoom_levels)
-        try:
-            base_max = minmax["max"]
-        except KeyError:
-            base_max = max(self.zoom_levels)
-        try:
-            resampling_lower = baselevels["lower"]
-        except KeyError:
-            resampling_lower = "nearest"
-        try:
-            resampling_higher = baselevels["higher"]
-        except KeyError:
-            resampling_higher = "nearest"
+        for v in minmax.values():
+            if v < 0 or not isinstance(v, int):
+                raise MapcheteConfigError(
+                    "invalid baselevel zoom parameter given: %s" % (
+                        minmax.values()
+                    )
+                )
+        base_min = minmax["min"] if "min" in minmax else min(self.zoom_levels)
+        base_max = minmax["max"] if "max" in minmax else max(self.zoom_levels)
+        resampling_lower = (
+            baselevels["lower"] if "lower" in baselevels else "nearest"
+        )
+        resampling_higher = (
+            baselevels["higher"] if "higher" in baselevels else "nearest"
+        )
         return dict(
             zooms=range(base_min, base_max+1),
             lower=resampling_lower,
@@ -360,11 +348,17 @@ class MapcheteConfig(object):
         else:
             raise MapcheteConfigError(
                 "Configuration has to be a dictionary or a .mapchete file.")
+        # make sure old input_files parameter is converted correctly
+        if "input_files" in raw and "input" in raw:
+            raise MapcheteConfigError("Either 'input_files' or'input allowed")
+        elif "input_files" in raw:
+            warnings.warn(
+                "'input_files' is deprecated and will be replaced by 'input'"
+            )
+            raw["input"] = raw.pop("input_files")
         # check if mandatory parameters are provided
         for param in _MANDATORY_PARAMETERS:
-            try:
-                assert param in raw
-            except AssertionError:
+            if not param in raw:
                 raise MapcheteConfigError("%s parameter missing" % param)
         # pixelbuffer and metatiling
         raw["pixelbuffer"] = self._set_pixelbuffer(raw)
@@ -372,9 +366,7 @@ class MapcheteConfig(object):
         raw["metatiling"] = self._set_metatiling(raw)
         raw["output"]["metatiling"] = self._set_metatiling(
             raw["output"], default=raw["metatiling"])
-        try:
-            assert raw["metatiling"] >= raw["output"]["metatiling"]
-        except AssertionError:
+        if not raw["metatiling"] >= raw["output"]["metatiling"]:
             raise MapcheteConfigError(
                 "Process metatiles cannot be smaller than output metatiles.")
         # absolute output path
@@ -383,28 +375,30 @@ class MapcheteConfig(object):
                 config_dir, raw["output"]["path"]))
         )
         # determine input files
-        if raw["input_files"] == "from_command_line":
-            if self.mode in ["memory", "continue", "overwrite"]:
-                try:
-                    assert single_input_file
-                except AssertionError:
-                    raise MapcheteConfigError(
-                        "please provide an input file via command line")
-                raw.update(input_files={"input_file": single_input_file})
+        if raw["input"] == "from_command_line" and (
+            self.mode in ["memory", "continue", "overwrite"]
+        ):
+            if not single_input_file:
+                raise MapcheteConfigError(
+                    "please provide an input file via command line")
+            else:
+                raw.update(input={"input_file": single_input_file})
+
         # return parsed configuration
         return raw, mapchete_file, config_dir
 
     def _set_pixelbuffer(self, config_dict):
         if "pixelbuffer" in config_dict:
-            assert isinstance(config_dict["pixelbuffer"], int)
-            assert config_dict["pixelbuffer"] >= 0
+            if not isinstance(config_dict["pixelbuffer"], int) and (
+                config_dict["pixelbuffer"] < 0
+            ):
+                raise ValueError("pixelbuffer must be an integer > 0")
             return config_dict["pixelbuffer"]
         else:
             return 0
 
     def _set_metatiling(self, config_dict, default=1):
         if "metatiling" in config_dict:
-            assert config_dict["metatiling"] in [1, 2, 4, 8, 16]
             return config_dict["metatiling"]
         else:
             return default
@@ -417,18 +411,18 @@ class MapcheteConfig(object):
         respective InputData class.
         """
         params = {}
-        input_files = {}
+        input_ = {}
         for name, element in self.raw.iteritems():
             if name not in _RESERVED_PARAMETERS:
                 out_element = self._element_at_zoom(name, element, zoom)
                 if out_element is not None:
                     params[name] = out_element
-            if name == "input_files":
-                input_files, process_area = input_files_at_zoom(
+            if name == "input":
+                input_, process_area = input_at_zoom(
                     self, name, element, zoom
                 )
         params.update(
-            input_files=input_files, output=self.output,
+            input=input_, output=self.output,
             process_area=process_area)
         return params
 
@@ -443,17 +437,16 @@ class MapcheteConfig(object):
         # overwrite if bounds are provided explicitly
         if user_bounds:
             # validate bounds
-            try:
-                assert len(user_bounds) == 4
-            except Exception:
+            if len(user_bounds) == 4:
+                bounds = user_bounds
+            else:
                 raise MapcheteConfigError("Invalid number of process bounds.")
-            bounds = user_bounds
 
-        input_files_bbox = self.at_zoom(zoom)["process_area"]
+        input_bbox = self.at_zoom(zoom)["process_area"]
         if bounds:
-            return box(*bounds).intersection(input_files_bbox)
+            return box(*bounds).intersection(input_bbox)
         else:
-            return input_files_bbox
+            return input_bbox
 
     def _element_at_zoom(self, name, element, zoom):
         """
@@ -473,18 +466,19 @@ class MapcheteConfig(object):
         """
         # If element is a dictionary, analyze subitems.
         if isinstance(element, dict):
-            sub_elements = element
+            if "format" in element:
+                return element
             out_elements = {}
-            for sub_name, sub_element in sub_elements.iteritems():
+            for sub_name, sub_element in element.iteritems():
                 out_element = self._element_at_zoom(
                     sub_name, sub_element, zoom)
-                if name == "input_files":
+                if name == "input":
                     out_elements[sub_name] = out_element
                 elif out_element is not None:
                     out_elements[sub_name] = out_element
             # If there is only one subelement, collapse unless it is
-            # input_files. In such case, return a dictionary.
-            if len(out_elements) == 1 and name != "input_files":
+            # input. In such case, return a dictionary.
+            if len(out_elements) == 1 and name != "input":
                 return out_elements.itervalues().next()
             # If subelement is empty, return None
             if len(out_elements) == 0: