+++ CHANGELOG.rst
@@ -3,6 +3,13 @@ Changelog
 #########
 
 ---
+0.6
+---
+* ``input_files`` config option now raises a deprecation warning and will be replaced with ``input``
+* abstract ``input`` types are now available which is necessary for additional non-file based input drivers such as DB connections
+
+
+---
 0.5
 ---
 * introduced iterable input data groups